created super class.
